requirements.txt
python-dotenv==0.20.0
requests==2.26.0
flask==2.0.3
Flask-SocketIO==5.1.1
paho-mqtt==1.6.1